[Storage] use async version fs.stat() (#5161)
use async version fs.stat() instead of fs.stateSync() to improve performance

* Update fs shim in rollup config

* Move fs helpers to utils.node.ts

* Add util.promisify() shim for browser rollup

* Removing testing code

* Re-gen test baseline
