+++ sdk/storage/storage-file/recordings/browsers/directoryclient/recording_listhandles_should_work.json
@@ -2,7 +2,7 @@
  "recordings": [
   {
    "method": "PUT",
-   "url": "https://fakestorageaccount.file.core.windows.net/share156874149465507289",
+   "url": "https://fakestorageaccount.file.core.windows.net/share156882667578404262",
    "query": {
     "restype": "share"
    },
@@ -10,19 +10,19 @@
    "status": 201,
    "response": "",
    "responseHeaders": {
-    "date": "Tue, 17 Sep 2019 17:31:34 GMT",
-    "last-modified": "Tue, 17 Sep 2019 17:31:34 GMT",
+    "date": "Wed, 18 Sep 2019 17:11:15 GMT",
+    "last-modified": "Wed, 18 Sep 2019 17:11:15 GMT",
     "server": "Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0",
-    "etag": "\"0x8D73B94E30469F1\"",
-    "x-ms-request-id": "d2f2ae7e-c01a-006f-6c7d-6d98a2000000",
+    "etag": "\"0x8D73C5B36D337ED\"",
+    "x-ms-request-id": "819e95dd-401a-005c-3644-6ec18f000000",
     "x-ms-version": "2019-02-02",
-    "x-ms-client-request-id": "829bcb3e-d115-460f-a5fd-7c139b90cf94",
+    "x-ms-client-request-id": "6dc5f010-02a8-44f5-9bc9-665d793004b9",
     "content-length": "0"
    }
   },
   {
    "method": "PUT",
-   "url": "https://fakestorageaccount.file.core.windows.net/share156874149465507289/dir156874149501008806",
+   "url": "https://fakestorageaccount.file.core.windows.net/share156882667578404262/dir156882667625403191",
    "query": {
     "restype": "directory"
    },
@@ -30,27 +30,27 @@
    "status": 201,
    "response": "",
    "responseHeaders": {
-    "date": "Tue, 17 Sep 2019 17:31:34 GMT",
-    "x-ms-file-change-time": "2019-09-17T17:31:34.7770317Z",
+    "date": "Wed, 18 Sep 2019 17:11:15 GMT",
+    "x-ms-file-change-time": "2019-09-18T17:11:15.8686755Z",
     "x-ms-file-attributes": "Directory",
     "x-ms-file-id": "13835128424026341376",
     "x-ms-request-server-encrypted": "true",
-    "x-ms-file-creation-time": "2019-09-17T17:31:34.7770317Z",
+    "x-ms-file-creation-time": "2019-09-18T17:11:15.8686755Z",
     "x-ms-file-parent-id": "0",
     "x-ms-file-permission-key": "15246684120248489204*13496228697838683005",
-    "x-ms-client-request-id": "4cc2f1db-2416-4b34-bea7-d6588b7c1a06",
+    "x-ms-client-request-id": "1b50e6f7-df2e-436a-84be-feaaa5d83fd4",
     "content-length": "0",
-    "last-modified": "Tue, 17 Sep 2019 17:31:34 GMT",
+    "last-modified": "Wed, 18 Sep 2019 17:11:15 GMT",
     "server": "Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0",
-    "etag": "\"0x8D73B94E31D6FCD\"",
-    "x-ms-request-id": "d2f2ae84-c01a-006f-717d-6d98a2000000",
-    "x-ms-file-last-write-time": "2019-09-17T17:31:34.7770317Z",
+    "etag": "\"0x8D73C5B37008823\"",
+    "x-ms-request-id": "819e95e0-401a-005c-3844-6ec18f000000",
+    "x-ms-file-last-write-time": "2019-09-18T17:11:15.8686755Z",
     "x-ms-version": "2019-02-02"
    }
   },
   {
    "method": "GET",
-   "url": "https://fakestorageaccount.file.core.windows.net/share156830612406803206/dir156830612568607181",
+   "url": "https://fakestorageaccount.file.core.windows.net/share156882667578404262/dir156882667625403191",
    "query": {
     "comp": "listhandles"
    },
@@ -58,18 +58,18 @@
    "status": 200,
    "response": "<?xml version=\"1.0\" encoding=\"utf-8\"?><EnumerationResults><Entries /><NextMarker /></EnumerationResults>",
    "responseHeaders": {
-    "date": "Tue, 17 Sep 2019 17:31:34 GMT",
+    "date": "Wed, 18 Sep 2019 17:11:15 GMT",
     "server": "Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0",
     "transfer-encoding": "chunked",
     "content-type": "application/xml",
-    "x-ms-request-id": "d2f2ae88-c01a-006f-757d-6d98a2000000",
+    "x-ms-request-id": "819e95e2-401a-005c-3a44-6ec18f000000",
     "x-ms-version": "2019-02-02",
-    "x-ms-client-request-id": "736f8ef2-6561-4baa-9a73-d9bafce1768f"
+    "x-ms-client-request-id": "5a68f034-3017-46e3-b864-8424daf2844d"
    }
   },
   {
    "method": "DELETE",
-   "url": "https://fakestorageaccount.file.core.windows.net/share156874149465507289",
+   "url": "https://fakestorageaccount.file.core.windows.net/share156882667578404262",
    "query": {
     "restype": "share"
    },
@@ -77,17 +77,17 @@
    "status": 202,
    "response": "",
    "responseHeaders": {
-    "date": "Tue, 17 Sep 2019 17:31:35 GMT",
+    "date": "Wed, 18 Sep 2019 17:11:15 GMT",
     "server": "Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0",
-    "x-ms-request-id": "d2f2ae8e-c01a-006f-7a7d-6d98a2000000",
+    "x-ms-request-id": "819e95e4-401a-005c-3c44-6ec18f000000",
     "x-ms-version": "2019-02-02",
-    "x-ms-client-request-id": "4c9ae987-f5b9-4400-8673-4c9abb6fd8c2",
+    "x-ms-client-request-id": "90b3d464-3b2d-43c1-94ef-71314f1e62f1",
     "content-length": "0"
    }
   }
  ],
  "uniqueTestInfo": {
-  "share": "share156874149465507289",
-  "dir": "dir156874149501008806"
+  "share": "share156882667578404262",
+  "dir": "dir156882667625403191"
  }
 }
