+++ sdk/storage/storage-file/rollup.base.config.js
@@ -23,7 +23,7 @@ const depNames = Object.keys(pkg.dependencies);
 const production = process.env.NODE_ENV === "production";
 
 export function nodeConfig(test = false) {
-  const externalNodeBuiltins = ["@azure/core-http", "crypto", "fs", "events", "os", "stream"];
+  const externalNodeBuiltins = ["@azure/core-http", "crypto", "fs", "events", "os", "stream", "util"];
   const baseConfig = {
     input: "dist-esm/src/index.js",
     external: depNames.concat(externalNodeBuiltins),
@@ -99,13 +99,16 @@ export function browserConfig(test = false, production = false) {
       shim({
         dotenv: `export function config() { }`,
         fs: `
-          export function statSync() { }
+          export function stat() { }
           export function createReadStream() { }
           export function createWriteStream() { }
         `,
         os: `
           export const type = 1;
           export const release = 1;
+        `,
+        util: `
+          export function promisify() { }
         `
       }),
       nodeResolve({
